Add a recovery mechanism for ID overflowing the integer range
I've decided on a stop-the-world strategy when the id count reaches the
upper limit. The goal is to minimize the checks impact on the common
case as the overflow should not happen often.

Have one idea how to polymorphically overwrite static fields on a class,
therefore went with this design. Not grinding our test suites to a halt
had priority over proper code style in that regard. Feel free to change
 this in the future if you have a better solution in mind.