Simplify calls to _reenumerate
Firstly, this called for a DRYer refactoring and seams like a sensible
behaviour in the first place (eg. I was expecting the method to behave
as a bound method in the first place). I know JS has prototypal
inheritance and this migh not be the most sensible solution, yet I don't
expect much going on near the id functionality. It is already
over-engineered and will handle *a lot* more stress than we're expecting
the system to under in the foreseeable future, so I'm comfortable with
it.