Add serialization mechanism to the Tree
I though about sorting nodes in the beginning before deserializing and
then it occurred to me that there is no tree traversal that visits a
child before a parent, so that won't be a concern.

I'm still not entirely sure where to put the functions, figure they
should be in a repository with appropriate high-level interface with
having some choice over the persistence method via the factory. But
where should the repository live then?