Introduce Resources into the view (#45)
After a talk with Dawid I've decided to change URL resource name to
Link, since no one was calling it URL ;)

This may look crappy, but we're getting rid of the infamous bleble
interface without any loss of functionality (if we can talk about
functionality at this stage ;d)

I'm pondering how to handle Resource displaying responsibility - having
one huge function is a code smell, yet for Resources to know how to
display themselves would be a breach of boundaries. And many small
functions seam like an overkill and a burden. Anyway: for now going with
the simplest thing that works, will see later :D