Extract ResourceEditor component
This commit doesn't change anything functionally

The main component was getting slightly out of hand. This also allowed
me to establish cleaner boundaries between domain objects - the
ResourceEditor is unaware of any Tree (and so would be TreeDisplay after
extraction) and the main component couldn't care less about what exactly
resources are [aside from creating the initial tree]

Had some issues with synchronizing changes on the address filed of the
form, yet the key approach (https://stackoverflow.com/a/53313430/9134286)
yielded good results. Conceptually you can think of it as both:
- an array of ResourceEditors that happens to haven only one element
- the ResourceEditor being spun up for every node that has a potential
to be edited by the user (which right now is the selected node) and
discarded when not needed

I'd like to than my former employer for making processors that allow
this a bit wasteful, yet marvelously elegant design