Introduce Formik
This is a great foundation for doing reflective forms, where any
boilerplate would be increasingly painful. Formik reduces the necessary
code to a bare minimum by providing an almost DSL-like experience

Adding and edition can be though of 2 different ways of submitting the
form, which right now has to be haxed, though (as designated in the
comments) a PR with a proper machinery is on the way.

Although deletion can be though of another way of submitting the form, it
does not require access to any of the fields, therefore I've decided to
keep it outside of the submission mechanism. After it's called the
component is disposed of anyway [because there is no semantics in
editing a node that doesn't exist]

Formik tutorial / overview I'd recommend:
https://www.youtube.com/watch?v=yNiJkjEwmpw