Add effective resource awareness (#52)
Before the change from the user point of view there was virtually no
different between Link and Text, right now that changes. We'll
capitalize on it soon!

Previously we were doing some haxes around Formik multi-path submission
yet it occurred to me that this whole mess could be avoided if we simply
splitted the functionality. A so I did!

Furthermore I've added a TODO for extracting the delete functionality to
another tiny component. It's turtles all the way down

As for the liberal use of "as any" - I don't want to be stuck in TS
peculiarities at this very moment