Add safeguards for future Resource addition
When adding a new Resource correctly (see comments in the resources.ts file)
the type system demands an entry for every components that does
type-based pattern matching on the Resource discriminated union members

Inspiration:
https://medium.com/@fillopeter/pattern-matching-with-typescript-done-right-94049ddd671c#809c

Limitations:
- the solution  doesn't check for excessive declarations / code handling
 a non-existing resource
- casting, casting everywhere, but we'll deal with this later
- two additional sets with Resource addition (aside from actually adding
the Resource)

I'm accepting the tradeof as adding Resources shouldn't be that common
and there is a comment in place to facilitate doing that correctly.
To the best of my knowledge this is the best that could be done

It actually allowed to simplify many mechanisms either directly (moving
the statically reflected type properties to the core domain) or
indirectly (I've realised that initial Formik fields can be simply the
resource object itself).

As a side note: I'm having second thoughts about choosing TypeScript for
this, as it clearly lacks the expressive power to implement such
features properly